no flag
skip the flag so we don't have to have a forked celery too.

I think that the child process once spawned gets the parent sys.path ~329-333 in the same file. So loading it with the env shouldn't really do anything…

See https://github.com/celery/billiard/issues/28 for more details.